Revamp project ids, add ProjectId.
We now allow a project to have zero or more ids, instead of an optional RepoId
and optional SrcURL. It would be odd for a project to have more than one of the
same kind of id, but as long as it didn't conflict with other projects, there's
no particular problem with that.

We also resolve depends directly by ID because duh.

Lastly, we allow ProjectFinderPlugins to resolve projects by ID as well as by
root. So if someone asks for a project by ID that we've never heard of, we give
the finder plugins a chance to pony up.

This is useful for two situations: one, we can provide platform projects that
are resolved only by ID (like the JDKs), and two, we can resolve faux projects
for Maven and Ivy artifacts, which will allow them to fit in more nicely with
all the things that rely on the project dependency graph.