Omit incidental projects from by name lookup.
Now that we turn all depends into projects, we don't want to flood the project
namespace.

This is a stopgap measure. I think the right thing to do is eliminate by name
lookup entirely and institute a full fledged workspace abstraction. Then we can
restrict ourselves to projects specifically added to your workspace, which will
be more precise and more manageable.