A bunch of Codexery.
Implemented a summarize-def view which shows docs and sigs for a def and its
members. One can expand docs and navigate the members (to summaries thereof or
to their defining source file).

Much work is needed, but this is a big step in the direction I've been aiming
for since I started integrating Codex into Scaled.