Codex is now workspace-wide.
This is the first step toward converting many project services to workspace
services. In most cases, we want things to operate across all projects in our
workspace. If I'm looking up a type or opening a file, it could be in any of
the projects in my workspace, so that should be the default operation.