Added codex-rename-element.
This is barely half an ass. To properly support rename refactoring (in OO
languages), it needs to know about constructors, method overrides, and any
other circumstance where a name occurs in multiple places. I'm not totally sure
I want to even try to do that in a language agnostic way... blah.