Revamped summary mode to pop docs instead of expand.
Having the buffer jump around and then having the docs all jammed into the big
list is not as nice as the popup. It's true that there's no way now to see docs
if they're SOO BIG that they don't fit on a single screen, but you can always
just jump to the definition itself and read the docs from the source.