Revamped run-last-test.
Now it captures a closure and reinvokes that. Maybe I'll worry about the memory
leak this represents later, or maybe not.

This also makes run-test-at-point work with run-last-test, and that displays
the test results in a popup in the current buffer (woo!), which is handy.

Also changed run-test-at-point to be C-c C-t C-p. More mnemonic.