Revamp compiler errors -> visit list.
Every project maintains its own list of the latest compiler errors. When we
visit a buffer for a project, if we're not currently visiting some other kind
of thing, then visit that project's compiler errors.

When a compilation completes for the current project, overwrite whatever the
current visit list is with this project's compiler errors.

This feels more DWIM.