Further revamping of exec window geometry.
Now we have three ways to get geometry:

1. If the execution specifies a custom geometry for itself, that geometry is
used.
2. If a geometry is specified for all executions, that is used.
3. Otherwise the current window's geometry is used (meaning output will be sent
to the current window).

We no longer track the ExecWindow specially because we just recompute the
desired geometry for a given execution every time it is invoked (reasonably
cheap) and then look for a window with that geometry. If we find one, we use
it, otherwise we create one.

This means that if you specif the exact same geometry for two executions,
they'll end up using the same window, but that sounds more like a feature than
a bug to me. It would be weird if we created two windows with the exact same
geometry because one would perfectly occlude the other. Might as well just
share the one window at that point.