First pass at cached project metadata.
The plan is for projects to cache their most recently resolved metadata so that
when a project is first resolved, it starts with the cached metadata and then
replaces it with metadata read from actual project files (Maven, Gradle,
whatever) once that is known.

Previously we had relied on blocking parsing of project files to obtain enough
metadata to get things going, but that isn't going to scale to build tools like
SBT and Gradle which take multiple seconds to initialize themselves. So this
new world order will allow us to show the buffer and get right to editing and
then backfill project information once the build tool in question has gotten
its shit together.

It should also marginally speed up start times for the "fast" projects because
we'll be able to just read a couple of files and go, and let the other
resolution happen on a background thread.