Codex revamp.
We now have a global database of Codex indices and those can be accessed
without resolving their associated project. This lays some groundwork for
making Codex work more nicely with async resolved projects, and generally makes
it cheaper to aggregate the two dozen indices needed to resolve a codex request
because we don't need to resolve the projects associated with those indices.

Since the vast majority of indices are for dependencies, which can be indexed
once and then never change, this is a win.