Show summaries in a separate window (if configged).
This underscores a need for a better system to automatically manage "extra info
displays" because this is something that any number of services would benefit
from.

It would be great to be able to say "Use this window for all aux info" and/or
"Use this window for aux info with this tag, this one for this tag, and this
one for the rest", etc. Then test output, compiler output, codex summary info,
exec console output, and any other "stuff you want to see next to your code"
could just ask for a window with its prioritized list of tags and plug right
into the system, rather than doing a poor job of managing these configs
themselves.