Phase one of project reorg.
The goal is to no longer have Project subclasses, but rather project component
providers. So if we see a particular sentinel file, providers can add one or
more components to a project as a result.

This splits off "all the project files" into the Filer component (wish it were
called Files, but name conflicts alas), and the project source directories into
the "Sources" component.

I still need to split the name + ids + dependencies out into one or more
components and I want to consider a way that components could be added based on
the existence of other components (ideally reactively).

Then I need to revamp the ProjectFinderPlugin into a ProjectComponentPlugin
which will add components to a project during resolution.