Phase two of project reorg.
No more subclasses of Project. Factored more things out into components.
ProjectFinderPlugin goes away and is replaced by ProjectRootPlugin which
identifies candidate project roots, and ProjectResolverPlugin which adds
components to projects.

Presently the resolvers duplicate some of the work of the root plugins (mainly
looking for sentinel files). I may want the root plugins to supply a set of
tags to apply to the resolved project and have the resolvers predicate
themselves on those tags. That might also be useful for prioritizing project
components (it's now possible for multiple resolver plugins to provide the same
component and the last one, which is currently arbitrary, will win).