Fix project resolution bugs, use info windows.
If we have a root, we can always resolve a project for it. We were resolving by
paths when we were given a known root which was bogus and it was dropping the
module info in the process. Now everything works better and is simpler.

Test output goes to 'tests' tagged window, execution output to 'exec' tagged
window, Codex summary info to 'codex' tagged window. The old ad-hoc execution
window geometry support and codex summary window geometry support is removed.