Some Analyzer reorg.
Analyzer has to be buffer state, not a project component. It's possible that
different buffers in a single project will have different analyzers (they could
be implemented in different languages and each have separate LangAnalyzers, for
example).

Also split out analyzer notes by store. We were already doing this under the
hood in LangAnalyzer, but it's simpler to just have Project know that notes are
on a per-store/buffer basis and then ProjectMode can accommodate that reality.

We use a special visit list that visits only the notes in the current buffer
and if there are no notes in the current buffer, it will jump to the notes in
the next (alphabetically sorted) buffer that does have notes.

This also factors the LangClient stuff into a pseudo project component just to
keep the LSP code together instead of intertwingling some of it into Project.