Revamps to allow runTestAtPoint via Intel intel.
This adds Intel.Defn which is a denormalized version of Codex's Def, and
Intel.enclosers which gives us Defn objects for all enclosing code
elements (function, class, module, whatnot).

We can then use this information to figure out the name of the function that
encloses the point, and assuming the signature information is correct,
correctly skip over nested enclosing functions and get to the real test
function.

Next I will attempt to wire this up using the results of an LSP client document
symbols request. The best laid plans of mice and men...