Wired up LangClient.enclosers; had to make it async.
Of course LSP has two ways to return info on symbols in a file, one which sucks
a lot and one which sucks a little bit less. Naturally most language servers
return the suckier one, because the second one didn't exist when they created
their implementation. Meanwhile, language clients realized that the original
form was woefully insufficient and now we have this awesome double burden of
supporting two forms of the same data. Foresight was apparently in short supply
on the LSP design team.